Handle Peekaboo failure scenarios
Add Peekaboo job failure handling to submitter, poker and tracker.

Submitting jobs can fail in the submitter and needs communicating to the
correlator. The way chosen here is to send a peekaboo-job task without
job ID into the normal pipe as an error indicator. The poker hands it to
the tracker as normal and the tracker short circuits it into a
peekaboo-report with failed result.

Tracking errors in the tracker produce a similar "synthetic" failed
report.

Jobs that failed within Peekaboo are synthesized by the tracker with a
certain likelihood for testing purposes.

The correlated report is changed to present jobs in a flat list instead
of a hash with the job IDs as keys because now not all jobs may have job
IDs.