Carry archive ignore decision into report
Don't just swallow samples like archives at the submitter stage but
submit a report without a job ID containing the information that a
sample was ignored for inclusion into the report. A separate job state
(as opposed to task state) is added to distinguish reasons for missing
job IDs such as submit-failed or archive-ignored. These are deliberately
not just the result/reason values used in the report to make it clear
that we're communicating something different here that may eventually
lead to a mention in the report but may be used for something else as
well.