Add caching of reports
In case the same sample is submitted multiple times, a large number of
redundant jobs may be submitted and tracked. This change adds caching
functionality based on the sha256 of the root sample submitted through
the API.

The cache is populated by the correlator and inspected by a new karton
called cache-responder which sits between the API and the classifier.
Because additional parameters may influence the outcome, they are also
checked and have to match exactly for a report to be re-used.

If an older job from the cache matches all criteria it is placed into
redis as an exact copy but under the new job's root_uid for the API to
find immediately. The karton task is dropped so it doesn't even reach
the classifier.

Usage of the cache is made optional by changing headers so that samples
are no longer addressed to the cache responder but the classifier as
before.