Switch job reports to hash
There's currently no reason not to use a hash for all the job reports.
This way we have less keys in redis. So switch to holding all job
reports in a single hash called extractor.reports.

This change showcases where a cache inconsistency can happen in the
cache responder. So we add handling and a warning for that case as well.