Support deduplication
Submitting duplicate samples multiple times may lead to massive churn
throughout the pipe and cause lots of Peekaboo jobs to be spawned,
particularly if the samples are archives. Therefore we add a deduper
which inspects the sample stream for duplicates. In case a duplicate is
found, it is held back, speculating on the already running job(s) to
finish and produce a report for the cache-responder to exploit for
short-circuiting the response. To that end, delayed samples are handed
back to the deduper itself via a delay queue. A cut-off deadline for
delaying is added to the deduper to prevent samples from starving.

Currently the deduper indefinitely accumulates locking and timeout keys
in two hashes. This will ultimately need garbage-collecting which should
be straight-forward because both keys' end of usefulness can be detected
by a report appearing and the cut-off time expiring, respectively.

Since the deduper is a delaying Karton very similar to the poker, common
functionality is moved into a module common.py. It remains otherwise
unchanged.

Similar to the cache-responder, the deduper is implemented as an
optional component of the pipe by changing task headers based on a
switches in the api and cache-responder.