common: More efficiently handle delayed tasks
Sleeping for a fixed amount of time would cause a noticeable delay
before new tasks were noticed and processed. Also it added a cumulative
delay on top of raw processing time.

By going back to using the blocking dequeueing function of the backend
we can be woken up as soon as new tasks arrive. We can still detect if
the timeout expired due to no new tasks arriving. In that case we can
switch to the next delay queue and start processing (or get blocked
waiting) there. If processing of a task takes very long or tasks arrive
continuously, we switch delay queues explicitly after the timeout
expired. That may leave tasks unprocessed in the old delay queue under
high load which is no worse than leaving jobs starving in the next delay
queue though.