deduper: Simplify garbage collection
Recording job holds by criteria and hold age by job ID in redis keys is
screaming for inconsistencies as showcased by garbage collection
complexity. Instead we record the time at which a task has been delayed
because of a hold in the task payload and apply the cutoff to that upon
rechecks. This makes it a per-task delay which could potentially make the
last task to arrive wait for longer than before.

To mitigate that we also now send outselves a poke to check the blocking
job for completion so that holds are removed quicker than by the garbage
collector and no longer need to be routinely ignored by duplicate jobs
after the cutoff. This makes held jobs sail on through as soon as the
hold is removed from redis.

These mechanics also implicitly implement a burst mode where any new
identical sample sets the hold anew. This should lessen churn for
duplicates whose jobs always fail.

All this greatly simplifies garbage collection which will now only very
rarely need to delete keys for really orphaned holds.

One pice is that the poker now also needs to ignore deduper holds when
deciding if more jobs are coming.