setup: Turn into module
Turn our code into a python module with setuptools setup.py, version,
description and everything to make it installable and useable more
flexibly.

Files are moved into respective subdirectories based on name and
function (group). Entrypoint functions are added so they're callable as
entrypoints from setuptools wrappers. That allows us to drop the
shebangs which we take as an opportunity to add some minimal module
docstrings.