all: Add external configuration
Add mechanics to support configuration from a file. We use Karton's
infrastructure for parsing config files because it's there and works
well enough. It even provides configuration from environment variables.
The only niggle is that no underscores are allowed in section names
and they also have to be lowercase, so readability suffers somewhat.

To gain a default argument parser for specifying the path to the
configuration file we switch to the also built-in Karton main()
entrypoint. To avoid forking that entrypoint to inject our non-blocking
backend in the poker and deduper, we move that injection into the common
DelayingKarton class. As the argument parser uses the class doc string
as synopsis default, we add class doc strings wherever they're still
missing.

Incidentally we rename some classes to come in line with our (not
explicitly formulated) component naming style. We also add the version
class member so the --version command line parameter works. This also
makes our bind register more nicely (i.e. with more metadata) in Karton.