container: Use pre-built images
Now that we have automatically-built container images in the Github
container registry, we can use them in the compose file to quicken
deployment and evaluation by users. The dev compose file is modified to
still build them locally so developers can easily test local
modifications. We override the image name as well to prevent confusion
if users switch between the dev and default environments and have
locally-built images.