install: Add dependencies
We create a virtualenv using the virtualenv command. It is not
guaranteed to be installed, so we add it to the list of packages we
install.

The same is true for unpackers used by sflock to extract archives which
is in turn used by the karton archive extractor. So we add those as well
to make the archive extractor function fully.