deduper: Do not misdirect dedupe hold on clearing
When clearing a dedupe hold due to its age being beyond the cutoff time,
the deduper would send that hold on to the cache-responder by tagging it
as a deduped sample. This would cause the cache-responder to run into
the following error:

Traceback (most recent call last):
  File "site-packages/karton/core/karton.py", line 178, in internal_process
    self.process(self.current_task)
  File "site-packages/extractor/cache/responder.py", line 76, in process
    sample = task.get_resource("sample")
  File "site-packages/karton/core/task.py", line 379, in get_resource
    raise TypeError("Resource was expected but not found")
TypeError: Resource was expected but not found

While not causing any functional problems by being redundant anyway it
certainly was inefficient and annoying. The respective logic was a
refactoring remnant from an earlier development state where we did not
have a separate dedupe hold task but instead tagged every single held
sample with the time it got held at and needed to pass it on once that
time aged beyond the cut-off threshold. Therefore we can simply drop
that passing-on of the task.