Rename to expander
Rename the project from extractor to expander, mainly because that name
is not in use yet as package name on PyPI. It also nicely captures the
"contracting" component of eventual correlation of individual analysis
results into an overal verdict for multiple samples e.g. generated from
archives.

Also move the project to a official repository in Github organisation
science-computing.