correlator: Switch to new graceful shutdown handling
Karton 5.0.0 also reworked shutdown handling, particularly the
GracefulKiller class into a graceful_killer() context handler. This
allows us to more elegantly disable that gracefull killer by just
overriding the context handler to do nothing. We still need this in the
correlator because we spawn two consumers there of which only the main
one is to install a signal handler. The other never shuts down during
processing but terminates after one run is completed.